Fixed handling of None values for limit, marker and is_public when listing images in client v1.
Currently, when images.list() is called in the v1 client, specifying
None for limit, marker or is_public results in an invalid value being
added to the query string, which results in an Invalid parameter error
from the server. This patch fixes this issue, making the behaviour
similar to other clients.

Change-Id: I91bb41109d1b19705c5673be15c86046e53b80ed